+++ SOURCE/main.c
@@ -12,8 +12,8 @@ int main(void)
 	int page = HOMEPAGE;
 	char user[16] = "GUEST";				//初始化用户名数组，系统限制，用户名仅为大写字母和数字，8位以内
 	FILE* fp;
-	short save_num = 1;
-	short mode = 0;//对战模式
+	int save_num = 1;
+	int mode = 0;//对战模式
 	SetSVGA64k(); 					//分辨率1024*768
 	//Initmouse(1, 1023, 1, 767); 
 	Initmouse(0, 1024, 0, 768);
